% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelle_viz.R
\name{annotation_chord_interactive}
\alias{annotation_chord_interactive}
\title{Title}
\usage{
annotation_chord_interactive(sce, colData1, colData2, color_by)
}
\arguments{
\item{color_by}{}
}
\description{
Title
}
